requirements.txt
flask
google-generativeai
pypdf
requests
beautifulsoup4
werkzeug